correctly implement Suspense for auth success page
